requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
numpy
pillow
pydantic
aiosqlite
python-multipart
jinja2
python-dotenv
aiofiles
